# src/main/java/com/simpleregisterlogin/dtos/EmailDTO.java
package com.simpleregisterlogin.dtos;

public class EmailDTO {

    private String email;

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}